% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_untb.R
\name{as.count}
\alias{as.count}
\alias{is.count}
\title{count}
\usage{
as.count(a, add = "")
}
\arguments{
\item{a}{species count data, see \link[untb]{count} documentation}

\item{add}{see \link[untb]{count} documentation}
}
\description{
Function from the \pkg{untb} package, see \link[untb]{count},
used for internal calculations in \link{fn.metaSIM}.
}
\details{
Used for internal calculations in \link{fn.metaSIM}.
}
\references{
Hankin, R. K. 2007. Introducing untb, an R package for simulating 
ecological drift under the unified neutral theory of biodiversity. Journal of 
Statistical Software 22:1--15. 
\url{http://cran.r-project.org/web/packages/untb/index.html}
}
\seealso{
\code{\link[untb]{count}}
}
