% Generated by roxygen2 (4.0.0): do not edit by hand
\name{data_family.rda}
\alias{data_family.rda}
\title{Zooplankton data including community composition, enviornmental variables
and geographic information}
\format{three data frames}
\source{
Manuscript
}
\description{
Zooplankton data including community composition, enviornmental variables
and geographic information
}

