% Generated by roxygen2 (4.0.0): do not edit by hand
\name{data_sim_parameters.rda}
\alias{data_sim_parameters.rda}
\title{Metacommunity simulation parameters representing ponds sampled in 2011 in
the Baltimore Ecosystem Study LTER}
\format{vectors and scalar variables used as simulation parameters}
\source{
Manuscript
}
\description{
Metacommunity simulation parameters representing ponds sampled in 2011 in
the Baltimore Ecosystem Study LTER
}

