% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fisher.ecosystem}
\alias{fisher.ecosystem}
\title{fisher.ecosystem}
\usage{
fisher.ecosystem(N, S, nmax, alpha = 2, c = 0)
}
\arguments{
\item{N}{Number of individuals}

\item{S}{Number of species}

\item{nmax}{Max number of species abundance classes}

\item{alpha}{Fisher's alpha. If not supplied it will be calculated from \code{N} and \code{S}}

\item{c}{Rare species advantage term}
}
\description{
Function from the \pkg{untb} package,
see \link[untb]{fisher.ecosystem}  -- used for internal
calculations in \link{fn.metaSIM}.
}
\details{
fisher.ecosystem
}
\references{
Hankin, R. K. 2007. Introducing untb, an R package for simulating
ecological drift under the unified neutral theory of biodiversity. Journal of
Statistical Software 22:1--15.
\url{http://cran.r-project.org/web/packages/untb/index.html}
}
\seealso{
\code{\link[untb]{fisher.ecosystem}}
}

