% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.combine.sim.csv.metadata}
\alias{fn.combine.sim.csv.metadata}
\title{fn.combine.sim.csv.metadata}
\usage{
fn.combine.sim.csv.metadata(out.dir.name = out.dir.name,
  file.name = "sim.metadata_", n.sims = nrow(data.simparameters))
}
\arguments{
\item{out.dir.name}{Name of the target directory in which simulation
results are saved, including the .csv metadata files associated with each simulation.}

\item{file.name}{currently must be \code{sim.metadata_}.  This is the file name used
by \code{fn.metaSIM} for the simulation metadata files.}

\item{n.sims}{The number of simulations saved to the target directory.}
}
\description{
Function from the \pkg{untb} package,
see \link[untb]{fisher.ecosystem}  -- used for internal
calculations in \link{fn.metaSIM}.
}
\details{
fn.combine.sim.csv.metadata
}

