% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.CqN.overlap}
\alias{fn.CqN.overlap}
\title{Community overlap of order q}
\usage{
fn.CqN.overlap(d, q.order = 1)
}
\arguments{
\item{d}{Site by species matrix or data frame of abundances or relative abundances}

\item{q.order}{Order q used to calculate overlap.  0 is presence/absence (more
sensitive to rare taxa), 1 is unbiased, and 2 is more sensitive to dominant taxa.}
}
\description{
Function to calculate community overlap among two or more sites, as
described by Jost et al. in Ch 6 in Biological Diversity (2011).
}
\details{
fn.CqN.overlap
}
\references{
Jost, L., A. Chao, and R. L. Chazdon. 2011. Compositional similarity and
beta diversity. Pages 68--84 in A. E. Magurran and B. J. McGill, editors. Biological
Diversity: Frontiers in Measurement and Assessment. Oxford University Press, Oxford, UK.
}

