% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.div_for_sim.data}
\alias{fn.div_for_sim.data}
\title{Simulation diversity partitioning}
\usage{
fn.div_for_sim.data(i.sim, out.dir.name, managed = site.management)
}
\arguments{
\item{i.sim}{identify simulation scenario to read}

\item{out.dir.name}{identify directory from which to read simulation outcomes}

\item{managed}{A vector identifying sites in a metacommunity simulation as "managed"
or "unmanaged"}
}
\description{
Read metacommunity simulation outcomes from a specified directory, then
calculate diversity partitions and write output to a .csv file in the same directory.
}
\details{
fn.div_for_sim.data
}

