% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.diversity.partition}
\alias{fn.diversity.partition}
\title{Diversity partitioning}
\usage{
fn.diversity.partition(dat.comm, q.order = 1)
}
\arguments{
\item{dat.comm}{}

\item{q.order}{}
}
\description{
Function called by \link{fn.metaSIM} to calculate alpha,
multiplicative beta, and gamma diversity for a site by species matrix
based on Jost (2007) and Chao et al. (2012).  This function relies on
the \link[vegetarian]{d} function in the \pkg{vegetarian} package.
}
\details{
fn.diversity.partition
}
\references{
Chao, A., C. H. Chiu, and T. C. Hsieh. 2012. Proposing a resolution to debates
on diversity partitioning. Ecology 93:2037--2051.

Jost, L. 2007. Partitioning diversity into independent alpha and beta components.
Ecology 88:2427--2439.

see the \href{http://cran.r-project.org/web/packages/vegetarian/index.html}{vegetarian}
package
}
\seealso{
\link[vegetarian]{d} in the \pkg{vegetarian} package
}

