% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.lambda.R
\name{fn.lambda}
\alias{fn.lambda}
\title{Calculate recruitment bias based on environmental filtering}
\usage{
fn.lambda(trait.optimum, niche.breadth, Ef, Ef.specificity)
}
\arguments{
\item{trait.optimum}{Optimum value for a given species along an environmental 
gradient.  Gradients in \link{fn.metaSIM} are restricted to the range [0,1], so 
this value must be in the range [0,1].}

\item{niche.breadth}{Niche breadth around a species' trait optimum.  The value of sigma 
in Fig 1 in Gravel et al. (2006).}

\item{Ef}{Value of the environmental filter at the site for which lambda is being 
calculated.}

\item{Ef.specificity}{The selection specificity of the environmental filter.}
}
\description{
Used by \link{fn.metaSIM}.  Uses niche breadth and an 
environmental gradient to determine how to bias species selection 
weights based on environmental filtering dynamics.  See Gravel et al. 2006.
}
\details{
Used by \link{fn.metaSIM}.
}
\references{
Gravel, D., C. D. Canham, M. Beaudet, and C. Messier. 2006. Reconciling niche and 
neutrality: the continuum hypothesis. Ecology Letters 9:399--409.
}
