% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_MCSim_0_3_release.R
\name{fn.lottery.recruit}
\alias{fn.lottery.recruit}
\title{Lottery recruitment}
\usage{
fn.lottery.recruit(vect.recruitment.weights, vect.taxa.list, scalar.JL)
}
\arguments{
\item{vect.recruitment.weights}{A vector of recruitment weights.}

\item{vect.taxa.list}{A vector of characters (or character strings) that is a 
list of species names.}

\item{scalar.JL}{Size of the assemblage (i.e., number of individuals to recruit).}
}
\description{
Lottery recruitment function for a single assemblage used 
by \link{fn.metaSIM}.
}
\details{
Used by \link{fn.metaSIM}.
}

