% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.make.landscape}
\alias{fn.make.landscape}
\title{Make a square landscape for fn.metaSIM}
\usage{
fn.make.landscape(n.sites = 100, JL.scale = 1, Ef.scale = -1,
Ef.specificity = 0.2, Ef.specificity.scale = 1, IL.scale = 2,
IL.intensity = 5, JM = 1000, JL.min = 1, ...)
}
\arguments{
\item{n.sites}{Number of sites in the landscape.}

\item{JL.scale}{Numeric value describing spatial heterogeneity in \code{JL}.
See \link{fn.metaSIM}.}

\item{Ef.scale}{Numeric value describing spatial heterogeneity in \code{Ef}.
See \link{fn.metaSIM}.}

\item{Ef.specificity}{Specificity of environmental filter.  See \link{fn.metaSIM}.}

\item{Ef.specificity.scale}{Numeric value describing the spatial heterogeneity
in \code{Ef.specificity}.  Currently can only take on values of -1, 1, or NA.
See \link{fn.metaSIM}.}

\item{IL.scale}{Numeric value describing spatial heterogeneity in \code{IL}.
See \link{fn.metaSIM}.}

\item{IL.intensity}{Number of individuals in the immigrant pool.  This value,
along with \code{JL}, determines \code{m} at each site.}

\item{JM}{Total number of individuals in the metacommunity.}

\item{JL.min}{Minimum number of individuals an assemblage is allowed to have.}

\item{\dots}{other values that are passed on to internal functions.}
}
\description{
Function to make a square landscape for \link{fn.metaSIM}.
}
\details{
fn.make.landscape
}
\seealso{
\link{fn.metaSIM}
}

