% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.make.landscape3}
\alias{fn.make.landscape3}
\title{Make a customizable landscape for fn.metaSIM3}
\usage{
fn.make.landscape3(x = d.geo$x, y = d.geo$y,
  management = d.siteinfo$managed, area_m2 = d.env.raw$area_m,
  Ef = Ef.obs, Ef.specificity.type = 0, JM = JM.est, I.rate.m2 = 1,
  mgmt.JL.increase = 0.25)
}
\arguments{
\item{x}{vector of x coordinates for sites/ponds}

\item{y}{vector of y coordinates for sites/ponds}

\item{management}{vector indicating site/pond management, or another grouping variable}

\item{area_m2}{vector of pond surface areas (m2)}

\item{Ef}{vector of values for site/pond environmental filter}

\item{Ef.specificity.type}{NOT USED, set to 0 as a default, environmental filter
type/strength, positive numeric value, or 'H1', 'H2', or 'neutral'}

\item{JM}{total metacommunity size}

\item{I.rate.m2}{areal immigration rate}

\item{mgmt.JL.increase}{NOT USED, default is 1 (no effect), factor indicating the
percent increase in population size at managed sites.}
}
\description{
Function to make a customizable landscape for \link{fn.metaSIM3}.
}
\details{
fn.make.landscape3
}
\seealso{
\link{fn.metaSIM3}
}

