% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.plot.sim}
\alias{fn.plot.sim}
\title{Plots to visualize a simulation}
\usage{
fn.plot.sim(sim.result, dist.type = "horn",
plot.timestep = c(1, sim.result$n.timestep), ...)
}
\arguments{
\item{sim.result}{A}{result from \link{fn.metaSIM}.}

\item{dist.type}{Distance}{metric, from \link[vegan]{vegdist} used to
plot an ordination of community composition.}

\item{plot.timestep}{Time}{steps to include in the plot.  Default is to
plot the initial and final time steps. Must be a vector of no more than
2 integers.}

\item{\dots}{Other}{plotting parameters passed onto \link{plot} and \link[vegan]{ordiplot}.}
}
\description{
NOT CURRENTLY USED... A function to plot select parameters
and diversity outcomes for the first and last timesteps in a simulation.
}
\details{
fn.plot.sim
}

