% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_MCSim_0_4_1_9002.R
\name{fn.recruit.Jt}
\alias{fn.recruit.Jt}
\title{Recruitment in a metacommunity with context}
\usage{
fn.recruit.Jt(landscape.site.coords = landscape$dat[, c("x", "y")], nu = nu,
              SWM.slope = SWM.slope, J.t.minus.1 = J.t.minus.1, 
              taxa.list = taxa.list, traits.Ef = dat.gamma.t0$trait.Ef, 
              trait.Ef.sd = trait.Ef.sd, traits.dispersal = dat.gamma.t0$trait.dispersal, 
              landscape.m = landscape$dat$m, landscape.Ef = landscape$dat$Ef, 
              landscape.Ef.specificity = landscape$dat$Ef.specificity, 
              landscape.JL = landscape$dat$JL)
}
\arguments{
\item{nu}{Numeric, Hubbell's "speciation rate", but can be interpreted as the 
probability of the appearance of a novel species.  If set to 0, no novel taxa 
will appear during the simulation.}

\item{SWM.slope}{Slope of dispersal kernel, see \link{fn.metaSIM}.  This value 
is used for \code{w} in eqn. 6 from Gravel et al. (2006) to model dispersal 
limitation.}

\item{J.t.minus.1}{Species composition of sites in a metacommunity at previous 
time step.}

\item{taxa.list}{A vector of character strings of species' names.}

\item{traits.Ef}{A vector of species' trait scores, numeric.}

\item{trait.Ef.sd}{A scalar value representing species' niche widths (numeric, 
currently one value is used for all species).}

\item{traits.dispersal}{A vector of species' dispersal trait scores.}

\item{landscape.site.coords}{A data.frame with xy-coordinates for each site.  Each row is a 
site in the metacommunity landscape.}

\item{landscape.m}{A vector of values for \code{m} for each site in the landscape.}

\item{landscape.Ef}{A vector of environmental filter \code{Ef} values for each site 
in the landscape.}

\item{landscape.Ef.specificity}{Environmental specificity at each site (numeric).}

\item{landscape.JL}{A vector of assemblage sizes (positive integers) for all sites 
in the landscape.}
}
\description{
Internal function called by \link{fn.metaSIM} to calculate local 
recruitment pools for all sites in a metacommunity, and call \link{fn.lottery.recruit}
to create a new assemblages (J) at time (t) after accounting for metacommunity 
composition at time (t-1), dispersal dynamics, landscape topology, and environmental 
filtering.
}
\details{
fn.recruit.Jt
}
\seealso{
\link{fn.lottery.recruit}, \link{fn.make.landscape}, \link{fn.metaSIM}
}

