% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.set.regional.species.pool}
\alias{fn.set.regional.species.pool}
\title{Initialize species pool for fn.metaSIM}
\usage{
fn.set.regional.species.pool(n.timestep = n.timestep, nu = nu,
speciation.limit = NA, JM, alpha.fisher, trait.dispersal.median,
trait.dispersal.range, taxa.list.prefix = "spp")
}
\arguments{
\item{n.timestep}{Number of generations (time steps) in the simulation.}

\item{nu}{Scalar value representing the probability that a novel species
is recruited.  Hubbell's \dQuote{speciation rate}.}

\item{speciation.limit}{A limit to the number of novel species that can occur
in the simulation.}

\item{JM}{Total number of individuals in the metacommunity.}

\item{alpha.fisher}{Fisher's alpha used to initiate the simulation.}

\item{trait.dispersal.median}{Median value for species' dispersal trait values.}

\item{trait.dispersal.range}{Range in variation allowed for species' trait
dispersal values.}

\item{taxa.list.prefix}{A character string to use as a prefix in species' names.}
}
\description{
Initialize species pool for fn.metaSIM
}
\details{
fn.set.regional.species.pool
}

