% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.sim.metadata.archive4}
\alias{fn.sim.metadata.archive4}
\title{Archive your simulation}
\usage{
fn.sim.metadata.archive4(sim.result = sim.result, var.dir = "SIM_OUTPUT",
var.filetype = ".rda", save.sim = TRUE, q.order = q.order,
keep.timesteps = c(1, 10))
}
\arguments{
\item{sim.result}{An object produced by \link{fn.metaSIM}, metacommunity simulation
output.}

\item{var.dir}{Directory where simulation outputs will be saved.}

\item{var.filetype}{File type used to save simulation outputs.  Currently must
be .rda.}

\item{save.sim}{Boolean (\code{TRUE} or \code{FALSE}), should the simulation output
be saved as a .rda file?  If \code{FALSE}, only the metadata file from the simulation
will be saved.}

\item{q.order}{Order \code{q} to use to calculate diversity statistics.  Can be a
single integer or a vector of integers.}

\item{keep.timesteps}{Timesteps to report in the metadata file.}
}
\description{
Saves metadata to a .csv file in an output directory, default
is SIM_OUTPUT. Optionally saves the simulation output as a .rda file.
These files can be big.
}
\details{
fn.sim.metadata.archive4
}

