% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.true.dist}
\alias{fn.true.dist}
\title{Custom function to calculate true distances (pairwise beta-diversities)}
\usage{
fn.true.dist(d = dat.comm, q.order = 1)
}
\arguments{
\item{d}{Site by species data matrix or data frame.}

\item{q.order}{Order \code{q}}
}
\description{
Custom function to calculate true distances based on Jost 2006,
Jost 2007, and Chao et al. 2012.
}
\details{
fn.true.dist
}
\references{
Chao, A., C.-H. Chiu, and T. C. Hsieh. 2012. Proposing a resolution
to debates on diversity partitioning. Ecology 93:2037--2051.
}

