% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fn.variation.partition}
\alias{fn.variation.partition}
\title{Variation partitioning using dbRDA}
\usage{
fn.variation.partition(dat.comm, E, S, q.order = 1)
}
\arguments{
\item{dat.comm}{community matrix}

\item{E}{A data frame in which columns are environemntal vars -- currently, the
simulation only produces one environmental gradient.}

\item{S}{A data frame of spatial variables (i.e., PCNM eigenvectors produced
by \link[vegan]{pcnm} in the \pkg{vegan} package.}

\item{q.order}{Order q to use to calculate among site dissimilarities in community
composition.  See \link{fn.true.dist}.}
}
\description{
Variation partitioning, similar to \link[vegan]{varpart} in
the \pkg{vegan} package, but uses \link[vegan]{capscale} (dbRDA) to
calculate [a], [b], [c], and [d] components instead of RDA.  Current
implementation only calculates species dissimilarities (of any order q) based
on Jost 2006, Jost 2007, and Chao et al. 2012 formula for relative turnover
(pairwise "true" beta diversities).
}
\details{
fn.variation.partition
}
\references{
Borcard, D., and P. Legendre. 2002. All-scale spatial analysis of ecological data
by means of principal coordinates of neighbour matrices. Ecological Modelling
153:51--68.

Peres-Neto, P. R., P. Legendre, S. Dray, and D. Borcard. 2006. Variation
partitioning of species data matrices - estimation and comparison of fractions.
Ecology 87:2614--2625.
}

