% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-MCSim.R
\docType{package}
\name{MCSim}
\alias{MCSim}
\title{MCSim: A metacommunity simulation package}
\description{
The MCSim package provides functions necessary for running iterative, lottery-based metacommunity simulations
}
