% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.coenoclines.r
\name{plot.coenoclines}
\alias{plot.coenoclines}
\title{plot standardized dispersal kernel}
\usage{
\method{plot}{coenoclines}(
  sim.result = NULL,
  landscape = NULL,
  Ef = NULL,
  trait.Ef = NULL,
  trait.Ef.sd = NULL,
  plot.trait.Ef.sd.0s = FALSE,
  y.max.val = NULL,
  ...
)
}
\arguments{
\item{sim.result}{output from a simulation, default is NULL}

\item{landscape}{a landscape object, default is NULL}

\item{Ef}{numeric vector of environmental filter values for sites in a simulation result, default is NULL}

\item{trait.Ef}{numeric vector of niche positions of species in metacommunity, default is NULL}

\item{trait.Ef.sd}{numeric vector of niche breadths, default is NULL}

\item{plot.trait.Ef.sd.0s}{should species with niche-breadth of 0 be plotted? Default is FALSE}

\item{y.max.val}{(optional) max value for y-axis}
}
\description{
plot standardized dispersal kernel
}
