% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dot.plots.R
\name{plot.dot.plots}
\alias{plot.dot.plots}
\title{plot dot plots}
\usage{
\method{plot}{dot.plots}(sim.result, timesteps = NA_integer_, spp.label.threshold.RA = 0.4)
}
\arguments{
\item{sim.result}{output from a simulation}

\item{timesteps}{(array of integers) timesteps to plot. Default is "NA" and will auto select timesteps}

\item{spp.label.threshold.RA}{(numeric) the relative abundance (RA) cutoff for taxa to label on the plot. Devault is to label taxa with a max RA >= 0.40}
}
\description{
plot dot plots of selected timesteps from an MCSim simulation
}
\examples{
\dontrun{
plot.dot.plots(my_sim_result) 
}
}
