% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.standardized.disp.kernel.R
\name{plot.standardized.disp.kernel}
\alias{plot.standardized.disp.kernel}
\title{plot standardized dispersal kernel}
\usage{
\method{plot}{standardized.disp.kernel}(
  sim.result = NULL,
  w = NULL,
  landscape = NULL,
  distance.matrix = NULL,
  ...
)
}
\arguments{
\item{sim.result}{simulation output, default is NULL}

\item{w}{dispersal kernel slope, default is NULL}

\item{landscape}{a landscape object, default is NULL}

\item{distance.matrix}{a distance matrix, default is NULL}
}
\description{
plot standardized dispersal kernel. Requires either (1) a sim.result, (2) a landsca;e and dispersal kernel slope(s) (w), or (3) a distance.matrix and dispersal kernel slope(s) (w).
}
