% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot_plots.R
\name{plot_dot_plots}
\alias{plot_dot_plots}
\title{plot dot plots}
\usage{
plot_dot_plots(
  sim_result,
  timesteps = NA_integer_,
  spp_label_threshold_RA = 0.4
)
}
\arguments{
\item{sim_result}{output from a simulation}

\item{timesteps}{(array of integers) timesteps to plot. Default is "NA" and will auto select timesteps}

\item{spp_label_threshold_RA}{(numeric) the relative abundance (RA) cutoff for taxa to label on the plot. Devault is to label taxa with a max RA >= 0.40}
}
\description{
plot dot plots of selected timesteps from an MCSim simulation
}
\examples{
\dontrun{
plot_dot_plots(my_sim_result) 
}
}
