% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_standardized_disp_kernel.R
\name{plot_standardized_disp_kernel}
\alias{plot_standardized_disp_kernel}
\title{plot standardized dispersal kernel}
\usage{
plot_standardized_disp_kernel(w = NULL, sim_result = NULL,
  landscape = NULL, distance_matrix = NULL, ...)
}
\arguments{
\item{w}{dispersal kernel slope, default is NULL}

\item{sim_result}{simulation output, default is NULL}

\item{landscape}{a landscape object, default is NULL}

\item{distance_matrix}{a distance matrix, default is NULL}
}
\description{
plot standardized dispersal kernel
}
