% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_MCSim_0_4_0_9000.R
\name{fn.make.landscape}
\alias{fn.make.landscape}
\title{make a simulation landscape}
\usage{
fn.make.landscape(site.coords = data.frame(), dist.mat = data.frame(),
  site.info = data.frame(), JM = 1000, I.rate.m2 = 1, area.m2 = 1,
  Ef.specificity = 0, Ef = 0.5, guess.site.coords = FALSE,
  list.of.stuff = NA)
}
\description{
fn.make.landscape
}

