% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_MCSim_0_4_1_9001.R
\name{fn.set.regional.species.pool}
\alias{fn.set.regional.species.pool}
\title{Initialize species pool for fn.metaSIM}
\usage{
fn.set.regional.species.pool(alpha.fisher = 1)
fn.set.regional.species.pool(gamma.abund = c(.9, .4, .2, .3))
}
\arguments{
\item{alpha.fisher}{Fisher's alpha used to initiate the simulation.}

\item{n.timestep}{Number of generations (time steps) in the simulation, can affect number of novel species. Default is 0.}

\item{nu}{Scalar value representing the probability that a novel species 
is recruited.  Hubbell's \dQuote{speciation rate}.}

\item{speciation.limit}{A limit to the number of novel species that can occur 
in the simulation, default is 0.}

\item{JM}{Total number of individuals in the metacommunity.}

\item{gamma.abund}{Vector of relative abundances or counts used to set regional species pool RAs}

\item{trait.dispersal}{Vector of trait scores for dispersal, should be in range 0 to 1, or NULL}

\item{trait.dispersal.median}{Median value for species' dispersal trait values.}

\item{trait.dispersal.range}{Range in variation allowed for species' trait 
dispersal values.}

\item{Ef.min}{Min limit for environmental gradient}

\item{Ef.max}{Max limit for environmental gradient}

\item{trait.Ef}{Vector or value of niche positions for species in regional pool}

\item{trait.Ef.sd}{Vector or value of niche breadths}

\item{taxa.list.prefix}{A character string to use as a prefix in species' names.}

\item{taxa.list.prefix}{Text appended as prefix to species names to prevent errors for column names when site by species matrices are created. Default is "spp".}
}
\description{
Initialize species pool for fn.metaSIM
}
\details{
fn.set.regional.species.pool
}

