% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasim.disturbance.R
\name{metasim.disturbance}
\alias{metasim.disturbance}
\alias{fn.metaSIM.disturbance}
\title{A metacommunity simulation for ecologists}
\usage{
metasim.disturbance(scenario.name, landscape.list, time.interval.durations)
}
\arguments{
\item{scenario.name}{text string naming the scenario}

\item{landscape.list}{list of landscapes to run in series}

\item{time.interval.durations}{vector of integers representing the number of timesteps to use for each landscape in the landscape_list, MUST be same length as landscape_list}

\item{J.t0.initial}{Initial species abundance matrix for timestep "t0"}

\item{scenario_name}{DEPRECATED: use \code{scenario.name}}

\item{landscape_list}{DEPRECATED: use \code{landscape.list}}

\item{time_interval_durations}{DEPRECATED: use \code{time.interval.durations}}

\item{J.t0_initial}{Deprecated: use \code{J.t0.initial}}

\item{...}{parameters sent to metasim()}
}
\description{
Wrapper function for MCSim::metasim, sends landscape list and interval duration list, along with other metasim parameters that will be fixed across all simulations
}
\examples{
\dontrun{
set.seed(1234) #set random seed

# make a landscapes
my.landscape.1 <- make.landscape(
  site.coords = data.frame(
    x = c(1, 2, 3, 4, 5),
    y = c(1, 3, 1, 5, 2)),
  Ef = c(-.8, -.6, 0, .25, .9),
  m = 0.5,
  JM = 10000)

my.landscape.2 <- make.landscape(
  site.coords = data.frame(
    x = c(1, 2, 3, 4, 5),
    y = c(1, 3, 1, 5, 2)),
  Ef = c(-.8, -.6, 0, 0, 0),
  m = 0.5,
  JM = 10000)

# niche positions, niche breadths, and relative abundances for three species
niche.positions <-  c(-.5, 0, .5)
niche.breadths <- c(.2, .2, 5)
regional.rel.abund <- c(.8, .1, .1)

sim.result <- metasim.disturbance(
  scenario.name = "niche_shift",
  landscape.list = list(my.landscape.1,my.landscape.2),
  time.interval.durations = c(10,10),
  trait.Ef = niche.positions,
  trait.Ef.sd = niche.breadths,
  gamma.abund = regional.rel.abund,
  W.r = 0,
  nu = 0.001,
  output.dir.path = "my_disturbance_sim_output_directory")
}
}
