% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasim.disturbance.R
\name{metasim.disturbance}
\alias{metasim.disturbance}
\alias{fn.metaSIM.disturbance}
\title{A metacommunity simulation for ecologists}
\usage{
metasim.disturbance(scenario_name, landscape_list, time_interval_durations)
}
\arguments{
\item{scenario_name}{text string naming the scenario}

\item{landscape_list}{list of landscapes to run in series}

\item{time_interval_durations}{vector of integers representing the number of timesteps to use for each landscape in the landscape_list, MUST be same length as landscape_list}

\item{...}{parameters sent to fn.metaSIM()}
}
\description{
Wrapper function for MCSim::fn.metaSIM, sends landscape list and interval duration list, along with other metaSim parameters that will be fixed across all simulations
}
